outsourced LabelLayoutUtil
